Move skipstage to controller
